Switch from num::bigint to ramp
commit 6d20033d9720c8e02cc68d997780ddf1fa23321d
Author: Hendrik Kunert <qkunerd@gmail.com>
Date:   Sun Jul 19 11:27:46 2015 +0200

    Cleanup after switchover to Ramp

commit 5a46e168532751c5e3143ccf9ec1eee150e02cc2
Author: Hendrik Kunert <qkunerd@gmail.com>
Date:   Fri Jul 17 14:19:59 2015 +0200

    Refactor PublicKey

commit 525926ab58c51be46af54821d2c6479b2125d639
Author: Hendrik Kunert <qkunerd@gmail.com>
Date:   Fri Jul 17 11:26:46 2015 +0200

    Refactor implementation of mod_inverse

commit 49782f2389d8c7fcaebf2d7b4e982306fb63ec33
Author: Hendrik Kunert <qkunerd@gmail.com>
Date:   Wed Jul 15 15:42:03 2015 +0200

    Add benchmark test for mod_pow

+++ src/key_pair_builder/key_pair.rs
@@ -1,6 +1,5 @@
-use num::bigint::BigInt;
+use ramp::Int;
 use bigint_extensions::ModPow;
-use num::traits::One;
 
 use public_key::PublicKey;
 use private_key::PrivateKey;
@@ -11,7 +10,7 @@ pub struct KeyPair {
 }
 
 impl KeyPair {
-    pub fn decrypt(&self, c: &BigInt) -> BigInt {
+    pub fn decrypt(&self, c: &Int) -> Int {
 
         let n = &self.public_key.n;
         let n_square = &self.public_key.n_squared;
@@ -19,6 +18,6 @@ impl KeyPair {
         let lambda = &self.private_key.lambda;
         let u = &self.private_key.denominator;
 
-        ((c.mod_pow(&lambda, &n_square) - BigInt::one()) / n * u) % n
+        ((c.mod_pow(&lambda, &n_square) - Int::one()) / n * u) % n
     }
 }
